consolidate the duplicated detect_versions function
